updated schema and domain classes
