+++ src/main/java/app/controller/CalculateTest.java
@@ -6,43 +6,43 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class CalculateTest {
-    public static void main(String[] args) {
-        // list of objects
-        List<Driver> drivers = new ArrayList<>();
-        List<Ticket> tickets = new ArrayList<>();
-        List<Accident> accidents = new ArrayList<>();
-
-        // creating objects to create vehicle
-        Agent agent = new Agent(1, "smith@gmail.com", "12345", "Smith", 1);
-        InsurancePlan insurancePlan = new InsurancePlan(1, agent, "2/8/2020", "5/10/2030",
-                            BigDecimal.valueOf(50.0));
-        Vehicle vehicle1 = new Vehicle(8858, 2010, 30.0f, 1, insurancePlan);
-
-        // creating objects to create driver
-        Driver driver1 = new Driver(1, "2020-01-01", 1);
-        drivers.add(driver1);
-        Driver driver2 = new Driver(2, "2021-08-08", 2);
-        drivers.add(driver2);
-        Driver driver3 = new Driver(3, "2014-02-08", 3);
-        drivers.add(driver3);
-
-        // creating ticket objects
-        Ticket ticket1 = new Ticket(1, 1, 1,"2023-01-08");
-        tickets.add(ticket1);
-        Ticket ticket2 = new Ticket(2, 2, 1,"2022-06-07");
-        tickets.add(ticket2);
-        Ticket ticket3 = new Ticket(3, 1, 1, "2021-05-08");
-        tickets.add(ticket3);
-
-        // creating accidents objects
-        Accident accident1 = new Accident(1, BigDecimal.valueOf(5000),1, 1);
-        accidents.add(accident1);
-        Accident accident2 = new Accident(2, BigDecimal.valueOf(7000), 1, 2);
-        accidents.add(accident2);
-        Accident accident3 = new Accident(3, BigDecimal.valueOf(2000), 1, 3);
-        accidents.add(accident3);
-
-        BigDecimal totalCost = Calculate.calculate(vehicle1,drivers, accidents, tickets);
-        System.out.println("Total cost for a year: " + totalCost);
-    }
+//    public static void main(String[] args) {
+//        // list of objects
+//        List<Driver> drivers = new ArrayList<>();
+//        List<Ticket> tickets = new ArrayList<>();
+//        List<Accident> accidents = new ArrayList<>();
+//
+//        // creating objects to create vehicle
+//        Agent agent = new Agent(1, "smith@gmail.com", "12345", "Smith", 1);
+//        InsurancePlan insurancePlan = new InsurancePlan(1, agent, "2/8/2020", "5/10/2030",
+//                            BigDecimal.valueOf(50.0));
+//        Vehicle vehicle1 = new Vehicle(8858, 2010, 30.0f, 1, insurancePlan);
+//
+//        // creating objects to create driver
+//        Driver driver1 = new Driver(1, "2020-01-01", 1);
+//        drivers.add(driver1);
+//        Driver driver2 = new Driver(2, "2021-08-08", 2);
+//        drivers.add(driver2);
+//        Driver driver3 = new Driver(3, "2014-02-08", 3);
+//        drivers.add(driver3);
+//
+//        // creating ticket objects
+//        Ticket ticket1 = new Ticket(1, 1, 1,"2023-01-08");
+//        tickets.add(ticket1);
+//        Ticket ticket2 = new Ticket(2, 2, 1,"2022-06-07");
+//        tickets.add(ticket2);
+//        Ticket ticket3 = new Ticket(3, 1, 1, "2021-05-08");
+//        tickets.add(ticket3);
+//
+//        // creating accidents objects
+//        Accident accident1 = new Accident(1, BigDecimal.valueOf(5000),1, 1);
+//        accidents.add(accident1);
+//        Accident accident2 = new Accident(2, BigDecimal.valueOf(7000), 1, 2);
+//        accidents.add(accident2);
+//        Accident accident3 = new Accident(3, BigDecimal.valueOf(2000), 1, 3);
+//        accidents.add(accident3);
+//
+//        BigDecimal totalCost = Calculate.calculate(vehicle1,drivers, accidents, tickets);
+//        System.out.println("Total cost for a year: " + totalCost);
+//    }
 }
