+++ src/main/resources/mappers/userMapper.xml
@@ -1,14 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "public-id" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="user">
-	
 	<resultMap id="result" type="user">
 		<result property="id" column="user_id"/>
-		<result property="email" column="email"/>
+		<result property="username" column="username"/>
 		<result property="password" column="password"/>
-		<result property="name" column="name"/>
 	</resultMap>
-	
 	<!-- ReadAll, Create, Read, Update, Delete -->
 	<select id="readAll" resultMap="result">
 		SELECT * FROM user;
@@ -17,10 +14,10 @@
 		SELECT * FROM user WHERE user_id = #{id};
 	</select>
 	<insert id="create" parameterType="User">
-		INSERT INTO user (user_id, email, password, name) VALUES (#{id}, #{email}, #{password}, #{name});
+		INSERT INTO user (user_id, username, password) VALUES (#{id}, #{username}, #{password});
 	</insert>
 	<update id = "update" parameterType="User">
-		UPDATE user SET email=#{email}, password=#{password}, name = #{name};
+		UPDATE user SET username=#{username}, password=#{password} WHERE user_id = #{id};
 	</update>
 	<delete id="delete" parameterType="int">
 		DELETE FROM user WHERE user_id=#{id};
