+++ src/main/resources/mybatis-config.xml
@@ -9,18 +9,14 @@
 			<transactionManager type="JDBC" />
 			<dataSource type="POOLED">
 				<property name="driver" value="com.mysql.cj.jdbc.Driver" />
-				<property name="url" value="${URL}" />
-				<property name="username" value="${USERNAME}" />
-				<property name="password" value="${PASSWORD}" />
+				<property name="url" value="${url}" />
+				<property name="username" value="${user}" />
+				<property name="password" value="${password}" />
 			</dataSource>
 		</environment>
 	</environments>
 	<mappers>
 		<mapper resource="mappers/userMapper.xml"/>
-		<mapper resource="mappers/accidentMapper.xml"/>
-		<mapper resource="mappers/ticketMapper.xml"/>
-		<mapper resource="mappers/vehicleMapper.xml"/>
-		<mapper resource="mappers/driverMapper.xml"/>
-		<mapper resource="mappers/customerMapper.xml"/>
+
 	</mappers>
 </configuration>
